+++ test/qr_detector_test.exs
@@ -22,7 +22,7 @@ defmodule Evision.QRCodeDetector.Test do
         type: :f32
       )
 
-    points_out = Evision.Nx.to_nx(points, Nx.BinaryBackend)
+    points_out = Evision.Mat.to_nx(points, Nx.BinaryBackend)
     assert 1 == Nx.to_number(Nx.all_close(points_nx, points_out))
 
     bin = Evision.Mat.to_binary(straight_qrcode)